Stop trying to share temp variables for phi nodes between cases. (#35)
Fixes #34

The correctness of the previous approach to handling phi nodes (the
point where the two haves of a disjunction meet, and variables that are
bound on both sides are considered to be bound by the disjunction) was
questionable. This is a much simpler approach whose correctness is
easier to see (because there is no sharing between cases of the phi
handling nodes).